tidy: Relocate substitution algorithm to package
